+++ inc/lilyParser/utils.ts
@@ -2,6 +2,9 @@
 const WHOLE_DURATION_MAGNITUDE = 128 * 3 * 5;
 
 
+const GRACE_DURATION_FACTOR = 0.2246;
+
+
 // Greatest common divisor & Least common multiple
 const gcd = (a: number, b: number): number => b === 0 ? a : gcd(b, a % b);
 /*const gcd = (a: number, b: number): number => {
@@ -76,6 +79,7 @@ const replaceSourceToken = (source: string, token: string): string => {
 
 export {
 	WHOLE_DURATION_MAGNITUDE,
+	GRACE_DURATION_FACTOR,
 	gcd,
 	lcm,
 	lcmMulti,